Organisations des lecteurs dans un module à part
- Organistation de tous les lecteurs de données dans un module à part.
- Création d'un nouveau sous-module utils
- Cleanup des import
