+++ rsesq-data/readers/__init__.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+"""
+Created on Sat Apr  8 13:56:17 2017
+@author: jnsebgosselin
+"""
+
+from read_mddelcc_cehq import MDDELCC_CEHQ_Reader
+from read_mddelcc_rses import MDDELCC_RSESQ_Reader
