+++ rsesq-data/readers/base.py
