+++ rsesq-data/readers/read_ec_hydat.py
